Refactor sections to utilize new Section component for improved layout consistency
- Replaced individual section headers in BlogSection, ContactSection, ProjectsSection, and SkillsSection with the new Section component for better structure and maintainability.
- Updated ProjectCard to use the Stars_Thinsmooth icon instead of the deprecated thinsmooth.svg file.
- Introduced a new font configuration in lib/font.tsx for consistent typography across components.
