+++ components/project.tsx
@@ -1,11 +1,11 @@
+import { Stars_Thinsmooth } from "@/components/icons";
+import { Skill } from "@/components/skill";
 import { Project } from "@/config/projects";
 import { cn } from "@/lib/utils";
 import { Component } from "@/type/component";
-import Image from "next/image";
 import Link from "next/link";
 import { useRef } from "react";
 import { useHover } from "usehooks-ts";
-import { Skill } from "./skill";
 
 export const ProjectCard: Component<Project> = ({ name, status, description, technologies, awards, url, githubUrl }) => {
   const ref = useRef<HTMLDivElement>(null);
@@ -46,12 +46,7 @@ export const ProjectCard: Component<Project> = ({ name, status, description, tec
               target="_blank"
               rel="noopener noreferrer"
             >
-              <Image
-                src={"/assets/thinsmooth.svg"}
-                alt="Award"
-                width={16}
-                height={16}
-              />
+              <Stars_Thinsmooth />
 
               <span className="text-xs">
                 Awarded
